+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Typescript support
     - The tasks support typescript source from now on
     - Introduce ```tslint``` task
+    - Introduce ```typescript``` option in the ```generate``` command
 - Introduce ```babel``` task
     - Use ```babel``` and ```babel:watch``` tasks to build your project
 - Use ```babel-runtime``` instead of ```babel-polyfill```